m_fix : empty arguments
