+++ src/exec/exec_main.c
@@ -6,7 +6,7 @@
 /*   By: ghan <ghan@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/10 16:22:10 by jun               #+#    #+#             */
-/*   Updated: 2021/10/06 11:14:59 by ghan             ###   ########.fr       */
+/*   Updated: 2021/10/10 20:49:08 by ghan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
